Delete old Business.ts file, rename BusinessV2 to Business.ts, update optional properties for Address, Identity, and Additional Questions
