+++ packages/webcomponents/src/api/Business.ts
@@ -24,16 +24,16 @@ export enum BusinessType {
 }
 
 export interface Address {
-  id: string;
-  platform_account_id: string;
-  line1: string;
-  line2: string;
-  postal_code: string;
-  city: string;
-  state: string;
-  country: string;
-  created_at: string;
-  updated_at: string;
+  id?: string;
+  platform_account_id?: string;
+  line1?: string;
+  line2?: string;
+  postal_code?: string;
+  city?: string;
+  state?: string;
+  country?: string;
+  created_at?: string;
+  updated_at?: string;
 }
 
 export interface Document {
@@ -53,22 +53,22 @@ export interface Document {
 }
 
 export interface Identity {
-  address: Address;
-  created_at: string;
-  dob_day: string;
-  dob_month: string;
-  dob_year: string;
-  documents: Document[];
-  email: string;
-  id: string;
-  is_owner: boolean;
-  metadata: any;
-  name: string;
-  phone: string;
-  platform_account_id: string;
-  ssn_last4: string;
-  title: string;
-  updated_at: string;
+  address?: Address;
+  created_at?: string;
+  dob_day?: string;
+  dob_month?: string;
+  dob_year?: string;
+  documents?: Document[];
+  email?: string;
+  id?: string;
+  is_owner?: boolean;
+  metadata?: any;
+  name?: string;
+  phone?: string;
+  platform_account_id?: string;
+  ssn_last4?: string;
+  title?: string;
+  updated_at?: string;
 }
 
 export interface ProductCategories {
@@ -79,13 +79,50 @@ export interface ProductCategories {
 }
 
 export interface AdditionalQuestions {
-  business_revenue: string;
-  business_payment_volume: string;
-  business_dispute_volume: string;
-  business_receivable_volume: string;
+  business_revenue?: string;
+  business_payment_volume?: string;
+  business_dispute_volume?: string;
+  business_receivable_volume?: string;
+}
+
+export interface ICoreBusinessInfo {
+  business_structure?: BusinessStructure;
+  business_type?: BusinessType;
+  legal_name?: string;
+  doing_business_as?: string;
+  industry?: string;
+  tax_id?: string;
+  website_url?: string;
+  email?: string;
+  phone?: string;
+}
+
+export class CoreBusinessInfo implements ICoreBusinessInfo {
+  public business_structure: BusinessStructure;
+  public business_type: BusinessType;
+  public legal_name: string;
+  public doing_business_as: string;
+  public industry: string;
+  public tax_id: string;
+  public website_url: string;
+  public email: string;
+  public phone: string;
+
+  constructor(coreBusinessInfo: ICoreBusinessInfo) {
+    this.business_structure = coreBusinessInfo.business_structure;
+    this.business_type = coreBusinessInfo.business_type;
+    this.legal_name = coreBusinessInfo.legal_name;
+    this.doing_business_as = coreBusinessInfo.doing_business_as;
+    this.industry = coreBusinessInfo.industry;
+    this.tax_id = coreBusinessInfo.tax_id;
+    this.website_url = coreBusinessInfo.website_url;
+    this.email = coreBusinessInfo.email;
+    this.phone = coreBusinessInfo.phone;
+  }
 }
 
 export interface IBusiness {
+  additional_questions: AdditionalQuestions | {};
   business_structure: BusinessStructure;
   business_type: BusinessType;
   bank_accounts: BankAccount[];
@@ -95,21 +132,21 @@ export interface IBusiness {
   email: string;
   id: string;
   industry: string;
-  legal_address: Address;
+  legal_address: Address | {};
   legal_name: string;
   metadata: any;
   owners: Identity[];
   phone: string;
   platform_account_id: string;
-  representative: Identity | null;
+  product_categories: ProductCategories;
+  representative: Identity | {};
   tax_id: string;
   updated_at: string;
   website_url: string;
-  product_categories: ProductCategories;
-  additional_questions: AdditionalQuestions;
 }
 
 export class Business implements IBusiness {
+  public additional_questions: AdditionalQuestions | {};
   public business_structure: BusinessStructure;
   public business_type: BusinessType;
   public bank_accounts: BankAccount[];
@@ -119,18 +156,17 @@ export class Business implements IBusiness {
   public email: string;
   public id: string;
   public industry: string;
-  public legal_address: Address;
+  public legal_address: Address | {};
   public legal_name: string;
   public metadata: any;
   public owners: Identity[];
   public phone: string;
   public platform_account_id: string;
-  public representative: Identity | null;
+  public representative: Identity | {};
   public tax_id: string;
   public updated_at: string;
   public website_url: string;
   public product_categories: ProductCategories;
-  public additional_questions: AdditionalQuestions;
 
   constructor(business: IBusiness) {
     this.business_structure = business.business_structure;
@@ -142,18 +178,20 @@ export class Business implements IBusiness {
     this.email = business.email;
     this.id = business.id;
     this.industry = business.industry;
-    this.legal_address = business.legal_address;
     this.legal_name = business.legal_name;
     this.metadata = business.metadata;
-    this.owners = business.owners;
     this.phone = business.phone;
     this.platform_account_id = business.platform_account_id;
-    this.representative = business.representative;
     this.tax_id = business.tax_id;
     this.updated_at = business.updated_at;
     this.website_url = business.website_url;
     this.product_categories = business.product_categories;
-    this.additional_questions = business.additional_questions;
+
+    // Form sections
+    this.legal_address = business.legal_address || { country: 'USA' };
+    this.representative = business.representative || {};
+    this.additional_questions = business.additional_questions || {};
+    this.owners = business.owners;
   }
 }
